Make CaptureLocationTaskData not a GeometryTaskData since we are storing the coordinates and not the geometry
